+++ CHANGELOG.md
@@ -9,6 +9,8 @@ MOSDL version history
 
 ### 0.2-SNAPSHOT (unreleased)
 - Not released yet.
+- New XSD generator for generating XML Schema files for data structures defined in an MO specification.
+- Rename of the Maven plugin configuration switch `mosdl-doc-type` to `doc-type`.
 
 ### Version 0.1.1 (2019-12-05)
 - #1: Maven plugin now properly creates target directory.